requirements.txt
paho-mqtt
sh
python-dotenv
